Move retain tests into the unit test module
